BUG: By clearing the entire "IndentConsistencyCopBufferCmds" augroup, pending autocmds for other buffers were deleted by an autocmd run in the current buffer
Now deleting only the buffer-local autocmds for the {event}s that fired.
Factored out s:InstallAutoCmd().
ENH: Added "check after write" feature, which triggers the IndentConsistencyCop whenever the buffer is written. To avoid blocking the user, in large buffers the check is only scheduled to run on the next 'CursorHold' event.